requirements.txt
Django
gunicorn
whitenoise
djangorestframework
django-debug-toolbar
psycopg2
requests